Split Name::Merge tests out of name_test.rb (#4708)
Moves the 5 tests exercising Name::Merge (merge_editors, merge_interests,
merge_orphans_log, merge_rolls_back_all_changes_if_a_step_raises,
merge_reassigns_misspelling_created_during_merge) into
test/models/name/merge_test.rb, matching app/models/name/merge.rb.

test_destroy_orphans_log stays behind - it exercises Location#destroy!'s
rss_log orphaning, not any Name::* module.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
